Added support for eth_listening. Moved net_version to same file as others
